pyproject.toml
[tool.poetry]
name = "beamer"
version = "1.0.2"
description = "Bridging rollups with L1 guaranteed security"
authors = ["Beamer Bridge Team <contact@beamerbridge.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://www.beamerbridge.com/"
repository = "https://github.com/beamer-bridge/beamer"
documentation = "https://docs.beamerbridge.com/"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
web3 = "^5.24.0"
click = "^8.0.3"
structlog = "^22.3.0"
python-statemachine = "^0.9.0"
prometheus-client = "^0.16.0"
lru-dict = "^1.1.7"
toml = "^0.10.2"
atomics = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^22.1"
isort = "^5.12.0"
flake8 = "^6.0.0"
flake8-tuple = "^0.4.1"
flake8-bugbear = "^23.2.13"
mypy = "^0.991"
eth-account = "^0.5.6"
eth-brownie = "^1.19.0"
pylint = "^2.16.2"
types-pkg-resources = "^0.1.3"
types-requests = "^2.28.11"
pdbpp = "^0.10.3"
pytest-cov = "^4.0.0"
shiv = "^1.0.3"
Sphinx = "^6.1.3"
furo = "^2022.12.7"
sphinxcontrib-mermaid = "^0.8"
sphinxcontrib-soliditydomain = {git = "https://github.com/beamer-bridge/sphinxcontrib-soliditydomain.git", rev = "main"}
types-toml = "^0.10.8"

[tool.poetry.scripts]
beamer = 'beamer.cli:main'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 99
include = '\.pyi?$'
extend-exclude = '''
/(
    \.git
  | \.mypy_cache
  | build
  | dist
)/
'''

[tool.isort]
line_length = 99
profile = "black"
multi_line_output = 3
ignore_whitespace = true

[[tool.mypy.overrides]]
module = ["brownie.*", "twisted.*"]
ignore_missing_imports = true

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true
check_untyped_defs = true
warn_unused_ignores = true
strict_equality = true
disallow_untyped_defs = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = ["beamer.tests.*"]
disallow_untyped_defs = false

[tool.pytest.ini_options]
filterwarnings = "ignore:Development network has a block height of"

[tool.coverage.run]
source = ["beamer"]
omit = ["beamer/tests/*"]
